Remove unneeded console.error
