+++ test/autoProcess/style.test.ts
@@ -40,54 +40,3 @@ STYLE_LANGS.forEach(([lang, ext]) => {
     });
   });
 });
-
-describe('style - postcss', () => {
-  const template = `<div></div><style>div{appearance:none;}</style>`;
-  const templateSass = `<div></div><style lang="scss">div{appearance:none;}</style>`;
-  const optsWithoutConfigFile = getAutoPreprocess({
-    postcss: {
-      plugins: [
-        require('autoprefixer')({
-          overrideBrowserslist: 'Safari >= 5.1',
-        }),
-      ],
-    },
-  });
-
-  it('should not transform plain css with postcss if { postcss: falsy }', async () => {
-    const preprocessed = await preprocess(template, getAutoPreprocess());
-    expect(preprocessed.toString()).not.toMatch(/-webkit-/);
-  });
-
-  it('should not transform plain css with postcss if { postcss: true } and no configuration file at cwd', async () => {
-    const preprocessed = await preprocess(
-      template,
-      getAutoPreprocess({
-        postcss: true,
-      }),
-    );
-    expect(preprocessed.toString()).not.toMatch(/-webkit-/);
-  });
-
-  it('should transform plain css with postcss if { postcss: { plugins... } }', async () => {
-    const preprocessed = await preprocess(template, optsWithoutConfigFile);
-    expect(preprocessed.toString()).toMatch(/-webkit-/);
-  });
-
-  it('should transform async preprocessed css with postcss if { postcss: { plugins... } }', async () => {
-    const preprocessed = await preprocess(templateSass, optsWithoutConfigFile);
-    expect(preprocessed.toString()).toMatch(/-webkit-/);
-  });
-
-  it('should transform plain css with postcss if { postcss: { configFilePath: ... } }', async () => {
-    const preprocessed = await preprocess(
-      template,
-      getAutoPreprocess({
-        postcss: {
-          configFilePath: './test/fixtures/',
-        },
-      }),
-    );
-    expect(preprocessed.toString()).toMatch(/-webkit-/);
-  });
-});
