+++ test/transformers/postcss.test.ts
@@ -4,6 +4,55 @@ import getAutoPreprocess from '../../src';
 import { preprocess } from '../utils';
 
 describe('transformer - postcss', () => {
+  const template = `<div></div><style>div{appearance:none;}</style>`;
+  const templateSass = `<div></div><style lang="scss">div{appearance:none;}</style>`;
+  const optsWithoutConfigFile = getAutoPreprocess({
+    postcss: {
+      plugins: [
+        require('autoprefixer')({
+          overrideBrowserslist: 'Safari >= 5.1',
+        }),
+      ],
+    },
+  });
+
+  it('should not transform plain css with postcss if { postcss: falsy }', async () => {
+    const preprocessed = await preprocess(template, getAutoPreprocess());
+    expect(preprocessed.toString()).not.toMatch(/-webkit-/);
+  });
+
+  it('should not transform plain css with postcss if { postcss: true } and no configuration file at cwd', async () => {
+    const preprocessed = await preprocess(
+      template,
+      getAutoPreprocess({
+        postcss: true,
+      }),
+    );
+    expect(preprocessed.toString()).not.toMatch(/-webkit-/);
+  });
+
+  it('should transform plain css with postcss if { postcss: { plugins... } }', async () => {
+    const preprocessed = await preprocess(template, optsWithoutConfigFile);
+    expect(preprocessed.toString()).toMatch(/-webkit-/);
+  });
+
+  it('should transform async preprocessed css with postcss if { postcss: { plugins... } }', async () => {
+    const preprocessed = await preprocess(templateSass, optsWithoutConfigFile);
+    expect(preprocessed.toString()).toMatch(/-webkit-/);
+  });
+
+  it('should transform plain css with postcss if { postcss: { configFilePath: ... } }', async () => {
+    const preprocessed = await preprocess(
+      template,
+      getAutoPreprocess({
+        postcss: {
+          configFilePath: './test/fixtures/',
+        },
+      }),
+    );
+    expect(preprocessed.toString()).toMatch(/-webkit-/);
+  });
+
   it('should return @imported files as dependencies', async () => {
     const template = `<style>@import './fixtures/style.css';</style>`;
     const opts = getAutoPreprocess({
